% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_comparison_spatial.R
\name{fig_comparison_spatial}
\alias{fig_comparison_spatial}
\title{Figure of spatial comparison between species-scale and network-scale cumulative effects assessments}
\usage{
fig_comparison_spatial()
}
\description{
Figure of spatial comparison between species-scale and network-scale cumulative effects assessments
}
