% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_comparaison_spatial_direct_indirect.R
\name{fig_comparison_spatial_direct_indirect}
\alias{fig_comparison_spatial_direct_indirect}
\title{Figure of spatial comparison between species-scale and network-scale cumulative effects assessments}
\usage{
fig_comparison_spatial_direct_indirect()
}
\description{
Figure of spatial comparison between species-scale and network-scale cumulative effects assessments
}
