% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_contribution.R
\name{fig_contribution}
\alias{fig_contribution}
\title{Figure of contribution of direct and indirect effects of stressors and of taxonomic groups to indirect effects of stressors to taxonomic groups}
\usage{
fig_contribution()
}
\description{
Figure of contribution of direct and indirect effects of stressors and of taxonomic groups to indirect effects of stressors to taxonomic groups
}
