% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_contribution_groups.R
\name{fig_contribution_groups}
\alias{fig_contribution_groups}
\title{Figure of contribution of direct and indirect effects of stressors to taxonomic groups}
\usage{
fig_contribution_groups()
}
\description{
Figure of contribution of direct and indirect effects of stressors to taxonomic groups
}
