% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_contribution_taxa.R
\name{fig_contribution_taxa}
\alias{fig_contribution_taxa}
\title{Figure of contribution of direct and indirect effects of stressors to individual taxa}
\usage{
fig_contribution_taxa()
}
\description{
Figure of contribution of direct and indirect effects of stressors to individual taxa
}
