% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_groups_spatial.R
\name{fig_groups_spatial}
\alias{fig_groups_spatial}
\title{Figure of the distribution of effects on selected species}
\usage{
fig_groups_spatial()
}
\description{
Figure of the distribution of effects on selected species
}
