% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_method.R, R/fig_method2.R
\name{fig_method}
\alias{fig_method}
\alias{fig_method2}
\title{Conceptual figure for methodology}
\usage{
fig_method()

fig_method2()
}
\description{
Conceptual figure for methodology

Conceptual figure for methodology
}
\section{Functions}{
\itemize{
\item \code{fig_method()}: metanetwork figure

\item \code{fig_method2()}: metanetwork figure

}}
