% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_species.R
\name{fig_species_spatial}
\alias{fig_species_spatial}
\title{Figure of the distribution of effects on selected species}
\usage{
fig_species_spatial()
}
\description{
Figure of the distribution of effects on selected species
}
