% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_stressors.R
\name{fig_stressors}
\alias{fig_stressors}
\title{Figure of contribution of direct and indirect effects of stressors to taxonomic groups}
\usage{
fig_stressors()
}
\description{
Figure of contribution of direct and indirect effects of stressors to taxonomic groups
}
