% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_anomalies.R
\name{out_anomalies}
\alias{out_anomalies}
\title{Script to extract anomalies between species-scale and network-scale cumualtive effects assessments}
\usage{
out_anomalies()
}
\description{
Script to extract anomalies between species-scale and network-scale cumualtive effects assessments
}
