% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_cea_km2.R,
%   R/out_interactions_simulations.R, R/out_mean_trophic_sensitivity.R
\name{out_cea_km2}
\alias{out_cea_km2}
\title{Script to extract cumulative effects scores / km2 for each species from the species and network-scale assessments}
\usage{
out_cea_km2()

out_cea_km2()

out_cea_km2()
}
\description{
Script to extract cumulative effects scores / km2 for each species from the species and network-scale assessments

Script to extract the mean trophic sensitivity per km2 for all species

Script to extract the mean trophic sensitivity per km2 for all species
}
