% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_cea_region.R
\name{out_cea_region}
\alias{out_cea_region}
\title{Script to extract mean network-scale cea in St. Lawrence System subregions}
\usage{
out_cea_region()
}
\description{
Script to extract mean network-scale cea in St. Lawrence System subregions
}
