% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_mechanisms.R
\name{out_mechanisms}
\alias{out_mechanisms}
\title{Script to extract correlations and anomalies between direct and indirect effects to assess amplification mechanisms}
\usage{
out_mechanisms()
}
\description{
Script to extract correlations and anomalies between direct and indirect effects to assess amplification mechanisms
}
