% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_plotEGSL.R
\name{plotEGSL}
\alias{plotEGSL}
\title{Function to create base plots of the St. Lawrence and the data contained in the package}
\usage{
plotEGSL(
  layers = "egslOutline",
  prj = slmetaPrj("default"),
  extent = c(-340251.1, 162300.3, 1018039.9, 990801.3),
  cols = NULL,
  borders = NULL,
  lwds = NULL,
  mar = c(0, 0, 0, 0),
  box = FALSE,
  axes = FALSE,
  background = NULL,
  graticules = NULL,
  scale = FALSE,
  northArrow = FALSE,
  prjText = FALSE,
  prjCol = "#000000",
  ...
)
}
\arguments{
\item{layers}{character, vector of layers to plot on the graph, in plotting order.}

\item{prj}{character, projection to use to plot the graph.}

\item{extent}{numeric(c(xmn, ymn, xmx, ymx)) extent to plot in prj units OR  | character(data) to use extent of data (i.e. 'canada').}

\item{cols}{character, vector of fill colors for all layers specified in layers param, in same order.}

\item{borders}{character, vector of border colors for all layers specified in layers param, in same order.}

\item{lwds}{numeric, vector of border widths for all layers specified in layers param, in same order.}

\item{mar}{numeric(c(0,0,0,0)), plot margins.}

\item{box}{logical, draw box around graph.}

\item{axes}{numeric(c(1,2,3,4)), draw axes around graph, 1 = bottom; 2 = left; 3 = top; 4 = right.}

\item{background}{character, graph background color.}

\item{graticules}{character, graph graticules color.}

\item{scale, }{logical, draw a scale bar (NOT YET IMPLEMENTED)}

\item{northArrow, }{logical, draw a north arrow}

\item{prjText, }{logical, add projection name to map}

\item{prjCol, }{character, projection name font color}
}
\value{
Opens a graphical interface with the plot
}
\description{
Function to create base plots of the St. Lawrence and the data contained in the package
}
\examples{
# Basic map
plotEGSL()

# Plot grid
plotEGSL(grid = T)

# Quebec
plotEGSL(layers = c('egslOutline','quebec','canada','usa'),
         cols = c('#2196a8', '#25364A','#25364A','#C7CBCE'),
         borders = '#f4f4f4',
         northArrow = T,
         mar = c(2,2,2,2),
         box = T,
         axes = 1:4,
         extent = 'quebec')

# Canada
plotEGSL(layers = c('egslOutline','canada','usa'),
        cols = c('#2196a8','#25364A','#C7CBCE'),
        borders = '#f4f4f4',
        northArrow = T,
        mar = c(2,2,2,2),
        box = T,
        axes = 1:4,
        extent = 'canada',
        prj = slmetaPrj('world'))

# Export
jpeg('./essai.jpg', width = 5, height = 5, res = 300, units = 'in')
plotEGSL(layers = 'egsl')
dev.off()
}
\keyword{lawrence}
\keyword{plot,}
\keyword{st.}
