% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_motifs.R
\name{plotMotifs}
\alias{plotMotifs}
\title{Function to plot motifs, from Beauchesne 2021}
\usage{
plotMotifs(
  motif,
  position = "",
  path = "",
  x = 0,
  y = 0,
  scalingX = 1,
  scalingY = 1,
  cex = 2,
  lwd = 2,
  lwd2 = 2,
  add = TRUE,
  posCol = "#000000",
  growth = "#AD524C",
  attack = "#AD524C",
  colLine = "#000000",
  conversion = "#53998e"
)
}
\description{
Function to plot motifs, from Beauchesne 2021
}
