% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_pubs.R
\name{render_ms}
\alias{render_ms}
\alias{render_ms_supmat}
\title{Publication rendering}
\usage{
render_ms(word = FALSE)

render_ms_supmat(word = FALSE)
}
\description{
Function to render publications from this project
}
\section{Functions}{
\itemize{
\item \code{render_ms()}: Render manuscript

\item \code{render_ms_supmat()}: Render manuscript supplementary materials

}}
