% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_metanetwork.R, R/fig_metanetwork_xl.R
\name{fig_metanetwork}
\alias{fig_metanetwork}
\alias{fig_metanetwork_xl}
\title{Export figures for metanetwork}
\usage{
fig_metanetwork()

fig_metanetwork_xl()
}
\description{
Export figures for metanetwork

Export figures for metanetwork
}
\section{Functions}{
\itemize{
\item \code{fig_metanetwork()}: metanetwork figure

\item \code{fig_metanetwork_xl()}: metanetwork figure

}}
